Refactor import statements in controllers to use lowercase model filenames
